Remove obsolete write_only_fields logic
(keeping in the form of simple versions, operating on hidden_fields only)
