--- src/api/response-types/write-only-field.ts
@@ -1,4 +1,0 @@
-export class WriteOnlyFieldType {
-  name: string;
-  is_set: boolean;
-}
