+++ src/components/index.ts
@@ -103,8 +103,8 @@ export { PermissionCategories } from './permission-categories';
 export { PermissionChipSelector } from './permission-chip-selector';
 export { PreviewRoles } from './preview-roles';
 export { PulpAboutModal } from './pulp-about-modal';
-export { PulpCopyButton } from './pulp-copy-button';
 export { PulpCodeBlock } from './pulp-code-block';
+export { PulpCopyButton } from './pulp-copy-button';
 export { PulpDownloadButton } from './pulp-download-button';
 export { PulpLabels } from './pulp-labels';
 export { PulpListToolbar } from './pulp-list-toolbar';
@@ -144,4 +144,3 @@ export { UploadSignatureModal } from './upload-signature-modal';
 export { UserForm } from './user-form';
 export { UserFormPage } from './user-form-page';
 export { WizardModal } from './wizard-modal';
-export { WriteOnlyField } from './write-only-field';
